Locate *major bug* in the CT_OUTOFPAGE3 method: B-phi incorrect as r^-2
The CT_FULL3D actually behaves without noticeable asymmetries, but
should be tested more carefully. It gives the correct 1/r profile of
B-phi. The bug in CT_OUTOFPAGE3 is still there, so don't use it until
it's fixed.

This commit also puts in additional measurements such as entropy, and
adds gamma_law_index as a data member to m2.